Revert removal of games.
All previously included games are included again, there is
obviously demand for almost all of them. I've received many
mails and bug reports to please keep a specific program.
So I've decided to re-enable all of them. In my opinion, it's
not sensible to switch to the alternative upstream at
https://sourceforge.net/projects/bsd-games/, because
many games have been removed from those sources.
This version of bsdgames is therefore essentially the same as
version 2.17-30.

Closes: #1069795, #1069807